quotactl02: flags for Q_XQUOTA{ON,OFF} is 32bit int
This fixes failure on ppc64 big endian systems:
  quotactl02.c:163: FAIL: quotactl() failed to turn off xfs quota and get xfs quota off status: EEXIST
  quotactl02.c:163: FAIL: quotactl() failed to turn on xfs quota and get xfs quota on status: EINVAL
  quotactl02.c:136: PASS: quoactl() succeeded to set and get xfs disk quota limits

+++ testcases/kernel/syscalls/quotactl/quotactl02.c
@@ -52,7 +52,7 @@ static struct fs_disk_quota set_dquota = {
 	.d_rtb_softlimit = 1000,
 	.d_fieldmask = FS_DQ_RTBSOFT
 };
-static unsigned short qflag = XFS_QUOTA_UDQ_ENFD;
+static uint32_t qflag = XFS_QUOTA_UDQ_ENFD;
 static const char mntpoint[] = "mnt_point";
 
 static struct t_case {
